+++ NEWS.md
@@ -1,5 +1,7 @@
 # crane 0.2.0.9015
 
+* Added `tbl_subgroups` and `g_forest` to make subgroup analysis and forest plots.
+
 * Added `list("assign_summary_type-arg:cat_threshold" = 0L)` to `theme_gtsummary_roche()`. Numeric variables with few levels will no longer default to summary type `'categorical'` in `gtsummary::tbl_summary()` and `tbl_roche_summary()`. (#79)
 
 * Added `tbl_null_report()` for cases without any information to display. (#11)